+++ src/OpenglWrapper.h
@@ -10,27 +10,27 @@ class OpenglWrapper // : GraphicsWrapper
 {
 public:
     
-    static void GenerateBuffers(size_t* ptr, size_t n = 1)
+    static void GenerateBuffers(size_t* ptr, int n = 1)
     {
 		glGenBuffers(n, (GLuint*)ptr);
     }
 
-    static void BindBuffer(size_t ID, size_t type = GL_ARRAY_BUFFER)
+    static void BindBuffer(size_t ID, int type = GL_ARRAY_BUFFER)
     {
 		glBindBuffer(type, (GLuint)ID);
     }
 
-    static void SetBufferData(size_t len, void* v, size_t type = GL_ARRAY_BUFFER, size_t draw_type = GL_STATIC_DRAW)
+    static void SetBufferData(size_t len, void* v, int type = GL_ARRAY_BUFFER, int draw_type = GL_STATIC_DRAW)
     {
 		glBufferData(type, len, v, draw_type);
     }
 
-    static void GenerateFrameBuffers(size_t* ptr, size_t n = 1)
+    static void GenerateFrameBuffers(size_t* ptr, int n = 1)
     {
         glGenFramebuffers(n, (GLuint*)ptr);
     }
 
-    static void BindFrameBuffer(size_t ID)
+    static void BindFrameBuffer(int ID)
     {
         glBindFramebuffer(GL_FRAMEBUFFER, ID);
     }
@@ -40,7 +40,7 @@ public:
         glBindFramebuffer(GL_READ_FRAMEBUFFER, 0);
     }
 
-    static void ReadBuffer(size_t buff = GL_COLOR_ATTACHMENT0)
+    static void ReadBuffer(int buff = GL_COLOR_ATTACHMENT0)
     {
         glReadBuffer(buff);
     }
@@ -55,12 +55,12 @@ public:
         glReadPixels(x, y, 1, 1, GL_RGBA, GL_UNSIGNED_INT, pixel);
     }
 
-    static void GenerateArrays(void* ptr, size_t n = 1)
+    static void GenerateArrays(void* ptr, int n = 1)
     {
 		glGenVertexArrays(n, (GLuint*)ptr);
     } 
 
-    static void BindArray(size_t ID)
+    static void BindArray(int ID)
     {
 		glBindVertexArray(ID);
     }
@@ -75,12 +75,12 @@ public:
 		return glCreateProgram();
 	}
 
-	static void LinkProgram(size_t program)
+	static void LinkProgram(int program)
 	{
 		glLinkProgram(program);
 	}
 
-	static void AttachShader(size_t program, int shader)
+	static void AttachShader(int program, int shader)
 	{
 		glAttachShader(program, shader);
 	}
@@ -95,12 +95,12 @@ public:
 		glCompileShader(shader);
 	}
 
-	static void GetShaderParam(int shader, size_t param, int* ptr)
+	static void GetShaderParam(int shader, int param, int* ptr)
 	{
 		glGetShaderiv(shader, param, ptr);
 	}
 
-    static void GetProgramParam(int program, size_t param, int* ptr)
+    static void GetProgramParam(int program, int param, int* ptr)
     {
 		glGetProgramiv(program, param, ptr);
     }
@@ -122,72 +122,72 @@ public:
 		glGetShaderInfoLog(shader, err_len, NULL, infoLog);
 	}
 
-	static void UseProgram(size_t shader)
+	static void UseProgram(int shader)
 	{
         glUseProgram(shader);
 	}
 
-	static void DeleteShader(size_t shader)
+	static void DeleteShader(int shader)
 	{
         glDeleteProgram(shader);
 	}
 
-    static void SetShaderInt(size_t ID, const char* name, int value)
+    static void SetShaderInt(int ID, const char* name, int value)
     {
 		glUniform1i(glGetUniformLocation(ID, name), (int)value);
     }
 
-    static void SetShaderFloat(size_t ID, const char* name, float value)
+    static void SetShaderFloat(int ID, const char* name, float value)
     {
 		glUniform1f(glGetUniformLocation(ID, name), value);
     }
 
-    static void SetShaderVec2(size_t ID, const char* name, float x, float y)
+    static void SetShaderVec2(int ID, const char* name, float x, float y)
     {
 		glUniform2f(glGetUniformLocation(ID, name), x, y);
     }
 
-    static void SetShaderVec3(size_t ID, const char* name, float x, float y, float z)
+    static void SetShaderVec3(int ID, const char* name, float x, float y, float z)
     {
 		glUniform3f(glGetUniformLocation(ID, name), x, y, z);
     }
 
-    static void SetShaderVec4(size_t ID, const char* name, float x, float y, float z, float w)
+    static void SetShaderVec4(int ID, const char* name, float x, float y, float z, float w)
     {
 		glUniform4f(glGetUniformLocation(ID, name), x, y, z, w);
     }
 
-    static void SetShaderMat2(size_t ID, const char* name, const float* m)
+    static void SetShaderMat2(int ID, const char* name, const float* m)
     {
 		glUniformMatrix2fv(glGetUniformLocation(ID, name), 1, GL_FALSE, m);
     }
 
-    static void SetShaderMat3(size_t ID, const char* name, const float* m)
+    static void SetShaderMat3(int ID, const char* name, const float* m)
     {
 		glUniformMatrix3fv(glGetUniformLocation(ID, name), 1, GL_FALSE, m);
     }
 
-    static void SetShaderMat4(size_t ID, const char* name, const float* m)
+    static void SetShaderMat4(int ID, const char* name, const float* m)
     {
 		glUniformMatrix4fv(glGetUniformLocation(ID, name), 1, GL_FALSE, m);
     }
 
-    static void EnableAttribute(size_t num)
+    static void EnableAttribute(int num)
     {
         glEnableVertexAttribArray(num);
     }
 
-    static void AttributePointer(int index, int size, size_t type, size_t stride, const void* ptr, bool normalize = GL_FALSE)
+    static void AttributePointer(int index, int size, int type, int stride, const void* ptr, bool normalize = GL_FALSE)
     {
         glVertexAttribPointer(index, size, type, normalize, stride, ptr);
     }
 
-    static void GenerateTextures(size_t* ID, size_t n = 1)
+    static void GenerateTextures(size_t* ID, int n = 1)
     {
         glGenTextures(n, (GLuint*)ID);
     }
 
-    static void BindTexture(size_t texture, size_t target = GL_TEXTURE_2D)
+    static void BindTexture(int texture, int target = GL_TEXTURE_2D)
     {
         glBindTexture(target, texture);
     }
@@ -197,12 +197,12 @@ public:
         BindTexture(0);
     }
 
-    static void ImageTexture(size_t format, size_t width, size_t height, unsigned char* data, size_t target = GL_TEXTURE_2D, size_t type = GL_UNSIGNED_BYTE, int lvl = 0, int border = 0)
+    static void ImageTexture(int format, int width, int height, unsigned char* data, int target = GL_TEXTURE_2D, int type = GL_UNSIGNED_BYTE, int lvl = 0, int border = 0)
     {
         glTexImage2D(target, lvl, format, width, height, border, format, type, data);
     }
 
-    static void ImageFrameBuffer(size_t ID)
+    static void ImageFrameBuffer(int ID)
     {
         glFramebufferTexture2D(GL_FRAMEBUFFER, GL_DEPTH_ATTACHMENT, GL_TEXTURE_2D, ID, 0);
     }
@@ -212,27 +212,27 @@ public:
         glGenerateMipmap(GL_TEXTURE_2D);
     }
 
-    static void TextureParameter(size_t pname, int param, size_t target = GL_TEXTURE_2D)
+    static void TextureParameter(int pname, int param, int target = GL_TEXTURE_2D)
     {
         glTexParameteri(target, pname, param);
     }
 
-    static void ActivateTexture(size_t texture = GL_TEXTURE0)
+    static void ActivateTexture(int texture = GL_TEXTURE0)
     {
         glActiveTexture(texture);
     }
 
-	static void DrawArrays(size_t count, size_t primitive = GL_TRIANGLES, size_t first = 0)
+	static void DrawArrays(int count, int primitive = GL_TRIANGLES, int first = 0)
 	{
         glDrawArrays(primitive, first, count);
 	}
 
-	static void DrawElements(size_t count, size_t primitive = GL_TRIANGLES, size_t type = GL_UNSIGNED_INT, const void* indices = nullptr)
+	static void DrawElements(int count, int primitive = GL_TRIANGLES, int type = GL_UNSIGNED_INT, const void* indices = nullptr)
 	{
         glDrawElements(primitive, count, type, indices);
 	}
 
-    static void DrawInstances(size_t count, size_t amount, size_t primitives = GL_TRIANGLES, size_t type = GL_UNSIGNED_INT, const void* indices = nullptr)
+    static void DrawInstances(int count, int amount, size_t primitives = GL_TRIANGLES, size_t type = GL_UNSIGNED_INT, const void* indices = nullptr)
     {
         glDrawElementsInstanced(GL_TRIANGLES, count, GL_UNSIGNED_INT, indices, amount);
     } 
@@ -252,7 +252,7 @@ public:
         glfwGetFramebufferSize(wind->getWindow(), W, H);
 	}
 
-	static void SetWindow(size_t W, size_t H)
+	static void SetWindow(int W, int H)
 	{
         glViewport(0, 0, W, H);
 	}
